perform EULA acceptance as it's own cli command
Introduce the `review-eulas` command on omg-cli to fetch/approve EULAs
from PivNet. This moves the acceptance logic from the `push-tiles`
command (baked into the ops_manager.Sdk) to its own command.

This command can be invoked with `--accept-all` in a CI environment.

The command is invoked on every run of ./deploy_pcf.sh. This isn't
strictly necessary, once a EULA for a given version has been accepted
then it does not need accepted again. We don't currently have a good
spot for an existing installation/change of tiles so doing this
unconditionally ensures we never run into EULA issues with updates.

changes:
- Replace all usages of `terraform-output-path` to `env-dir`. This fixes
  calling consistency with all of the omg-cli commands. Update the
  jumpbox to upload all config files specified in `config.ConfigFiles`.
  Update commands that load the config to use the config.*FromEnvDir
  versions.
- Rename env/<env name>/env.json to terraform_output.json
- Add PivnetApiToken to env/<env name>/config.json
- Add `review-eulas` command, related pivnet.Sdk API
- Add `omg-cli review-eulas` step to ci/tasks/push-tiles.sh
- Rename all pointer reciever names (eg dic* DeleteInstallationCommand) to
  cmd for consistency.
- Remove duplicated variable declaration in
  ci/tasks/create-infrastructure.sh

Change-Id: I06d0b35b23015adf275f92eb079805334924e20c