add cleanup-project command
This command attempts to clear out upgraded Ops Man VMs (eg via c0
pipeline) and abanonded BOSH created vms (eg a failed
delete-installation)

The findVMs call is complex because it's hard (or I can't figure out)
how to do advanced filtering with the AggregateListInstances call.

The other option would have been to blindly delete all VMs in a network
or project. I don't think this is appropriate because a project could
host multiple PCF deployments. Terraform should be the responsible party
for deleting resources it manages. BOSH should be the responsible party
for the same (though it can fail).

Change-Id: I96fc4c92222f3e86846132fd78bb6626159f2917