generate service accounts in terraform
Take advantage of the new service_account_key resource and move the
creation of (most) service accouts to the terraform.

We still create a single key for terraform to do it's job but this may
not be necessary.

Added some code in omg-cli to handle the base64 output from terraform.

Change-Id: Ic63a014259efe7f9f2cd51421b132f0c4e9034ca