util/delete_deployment: remove env dir on success
Clear out the directory after a successful deletion. The directory is
created by `deploy_pcf.sh` so it's fair to assume it should be cleaned
up by `delete_deployment.sh`.

This is good for users who make a configuration issue. Leaving around
the state/config files means they have generated values (eg in
config.json or terraform.tfvars) that are invalid or stale.

Change-Id: If03ddcefdfee683e031ff40453791c710d50763c