Try turning on go mod support manually
I am attempting to get Concourse to build. Things work for me locally,
and I believe the difference is: I'm downloading the project outside of
the GOPATH, while concourse is inside the GOPATH.

Therefore: I'm hoping turning on the GO111MODULE env flag will fix the
build issues on concourse.