Update dependencies
This change updates github.com/pivotal-cf/om to 0.44.0 to benefit from retry improvements in that package. It was not updated to the latest version (0.47.0) as that required a significant refactor to our client code.

All other dependencies were updated to their latest respective versions. This required modififying SCP code to omit "~" when specifying the destination file path.